% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbmcols.R
\name{cb.cols}
\alias{cb.cols}
\title{cb.cols}
\usage{
cb.cols(col, alpha = 1)
}
\arguments{
\item{col}{Which color you want to use (black, orange, sky blue, green,
yellow, blue, vermillion, pink)}

\item{alpha}{Color transparency}
}
\value{
col Hex code for desired color
}
\description{
Generate colors from a colorblind-friendly palette
}
\author{
Daniel B. McArtor (dmcartor@nd.edu) [aut, cre]
}

