% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbmcols.R
\name{col.options}
\alias{col.options}
\title{col.options}
\usage{
col.options(alpha = 1)
}
\arguments{
\item{alpha}{Color transparency}
}
\description{
Look at the output for the colorblind palettes available in this package
}
\author{
Daniel B. McArtor (dmcartor@nd.edu) [aut, cre]
}

