% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbmcols.R
\name{nd.cols}
\alias{nd.cols}
\title{nd.cols}
\usage{
nd.cols(col, alpha = 1)
}
\arguments{
\item{col}{Which color you want to use; either a string ('blue', 'gold'), or
number (1 = 'blue', 2 = 'gold')}

\item{alpha}{Color transparency}
}
\value{
col Hex code for desired color
}
\description{
Get Notre Dame blue and gold
}
\author{
Daniel B. McArtor (dmcartor@nd.edu) [aut, cre]
}

